+++ styleframe/style_frame.py
@@ -1,23 +1,25 @@
 import datetime as dt
-import numpy as np
 import pathlib
-import pandas as pd
 
-from functools import partial
-
-from .deprecations import deprecated_kwargs
-from . import utils
-from copy import deepcopy
 from collections import OrderedDict
 from collections.abc import Iterable
-from openpyxl import load_workbook
-from openpyxl.cell.cell import get_column_letter
-from openpyxl.xml.functions import fromstring, QName
+from copy import deepcopy
+from functools import partial
+from typing import Any, Dict, List, Optional, Set, Tuple, Union
+
+import numpy as np
+import pandas as pd
+
+from openpyxl import load_workbook, Workbook
+from openpyxl.cell.cell import get_column_letter, Cell
 from openpyxl.utils import cell
+from openpyxl.worksheet.worksheet import Worksheet
+from openpyxl.xml.functions import fromstring, QName
 
 from styleframe.container import Container
 from styleframe.series import Series
 from styleframe.styler import Styler, ColorScaleConditionalFormatRule
+from . import utils
 
 try:
     pd_timestamp = pd.Timestamp
@@ -25,7 +27,7 @@ except AttributeError:
     pd_timestamp = pd.tslib.Timestamp
 
 
-class StyleFrame(object):
+class StyleFrame:
     """
     A wrapper class that wraps a :class:`pandas.DataFrame` object and represent a stylized dataframe.
     Stores container objects that have values and styles that will be applied to excel
@@ -34,25 +36,31 @@ class StyleFrame(object):
           a list of dictionaries or another StyleFrame.
     :param styler_obj: Will be used as the default style of all cells.
     :type styler_obj: :class:`.Styler`
+    :param columns: Names of columns to use. Only applicable if ``obj`` is :class:`numpy.ndarray`
+    :type columns: None or list[str]
     """
-    P_FACTOR = 1.3
-    A_FACTOR = 13
+    P_FACTOR: Union[int, float] = 1.3
+    A_FACTOR: Union[int, float] = 13
 
-    def __init__(self, obj, styler_obj=None):
+    def __init__(self, obj, styler_obj: Optional[Styler] = None, columns: Optional[List[str]] = None):
         from_another_styleframe = False
         from_pandas_dataframe = False
         if styler_obj and not isinstance(styler_obj, Styler):
             raise TypeError('styler_obj must be {}, got {} instead.'.format(Styler.__name__, type(styler_obj).__name__))
-        if isinstance(obj, pd.DataFrame):
+        if isinstance(obj, (pd.DataFrame, np.ndarray)):
             from_pandas_dataframe = True
+            if isinstance(obj, np.ndarray):
+                obj = pd.DataFrame(obj)
+                if columns:
+                    obj = obj.rename(columns=dict(zip(obj.columns, columns)))
             if obj.empty:
                 self.data_df = deepcopy(obj)
             else:
-                self.data_df = obj.applymap(lambda x: Container(x, deepcopy(styler_obj)) if not isinstance(x, Container) else x)
+                self.data_df = obj.map(lambda x: Container(x, deepcopy(styler_obj)) if not isinstance(x, Container) else x)
         elif isinstance(obj, pd.Series):
             self.data_df = obj.apply(lambda x: Container(x, deepcopy(styler_obj)) if not isinstance(x, Container) else x)
         elif isinstance(obj, (dict, list)):
-            self.data_df = pd.DataFrame(obj).applymap(lambda x: Container(x, deepcopy(styler_obj)) if not isinstance(x, Container) else x)
+            self.data_df = pd.DataFrame(obj).map(lambda x: Container(x, deepcopy(styler_obj)) if not isinstance(x, Container) else x)
         elif isinstance(obj, StyleFrame):
             self.data_df = deepcopy(obj.data_df)
             from_another_styleframe = True
@@ -69,17 +77,21 @@ class StyleFrame(object):
         self._columns_width = obj._columns_width if from_another_styleframe else OrderedDict()
         self._rows_height = obj._rows_height if from_another_styleframe else OrderedDict()
         self._has_custom_headers_style = obj._has_custom_headers_style if from_another_styleframe else False
-        self._cond_formatting = []
+        self._cond_formatting: List[ColorScaleConditionalFormatRule] = []
         self._default_style = styler_obj or Styler()
         self._index_header_style = obj._index_header_style if from_another_styleframe else self._default_style
 
         self._known_attrs = {'at': self.data_df.at,
                              'loc': self.data_df.loc,
                              'iloc': self.data_df.iloc,
-                             'applymap': self.data_df.applymap,
                              'groupby': self.data_df.groupby,
                              'index': self.data_df.index,
-                             'fillna': self.data_df.fillna}
+                             'fillna': self.data_df.fillna,
+
+                             # applymap is deprecated in pandas > 2
+                             'applymap': self.data_df.map,
+                             'map': self.data_df.map
+        }
 
     def __str__(self):
         return str(self.data_df)
@@ -112,15 +124,15 @@ class StyleFrame(object):
             raise AttributeError("'{}' object has no attribute '{}'".format(type(self).__name__, attr))
 
     @property
-    def columns(self):
+    def columns(self) -> List[Container]:
         return self.data_df.columns
 
     @columns.setter
-    def columns(self, columns):
+    def columns(self, columns: Iterable) -> None:
         self.data_df.columns = [col if isinstance(col, Container) else Container(value=col)
                                 for col in columns]
 
-    def _get_column_as_letter(self, sheet, column_to_convert, startcol=0):
+    def _get_column_as_letter(self, sheet: Worksheet, column_to_convert, startcol: int = 0) -> str:
         col = column_to_convert.value if isinstance(column_to_convert, Container) else column_to_convert
         if not isinstance(col, (int, str)):
             raise TypeError("column must be an index, column letter or column name")
@@ -142,18 +154,14 @@ class StyleFrame(object):
         return column_as_letter
 
     @classmethod
-    @deprecated_kwargs(('sheetname',))
-    def read_excel(cls, path, sheet_name=0, read_style=False, use_openpyxl_styles=False,
-                   read_comments=False, **kwargs):
+    def read_excel(cls, path: str, sheet_name: Union[str, int] = 0, read_style: bool = False,
+                   use_openpyxl_styles: bool = False, read_comments: bool = False, **kwargs) -> 'StyleFrame':
         """
         Creates a StyleFrame object from an existing Excel.
 
         .. note:: :meth:`read_excel` also accepts all arguments that :func:`pandas.read_excel` accepts as kwargs.
 
         :param str path: The path to the Excel file to read.
-        :param sheetname:
-              .. deprecated:: 1.6
-                 Use ``sheet_name`` instead.
         :param sheet_name: The sheet name to read. If an integer is provided then it be used as a zero-based
                 sheet index. Default is 0.
         :type sheet_name: str or int
@@ -174,7 +182,7 @@ class StyleFrame(object):
         :rtype: :class:`StyleFrame`
         """
 
-        def _get_scheme_colors_from_excel(wb):
+        def _get_scheme_colors_from_excel(wb: Workbook) -> List[str]:
             xlmns = 'http://schemas.openxmlformats.org/drawingml/2006/main'
             if wb.loaded_theme is None:
                 return []
@@ -191,7 +199,7 @@ class StyleFrame(object):
                         colors.append(accent.attrib['val'])
             return colors
 
-        def _get_style_object(sheet, theme_colors, row, column):
+        def _get_style_object(sheet: Worksheet, theme_colors: List[str], row: int, column: int) -> Union[Cell, Styler]:
             cell = sheet.cell(row=row, column=column)
             if use_openpyxl_styles:
                 return cell
@@ -218,18 +226,17 @@ class StyleFrame(object):
             for col_index, col_name in enumerate(sf.columns):
                 col_index_in_excel = col_index + 1
                 if col_index_in_excel == excel_index_col:
-                    for row_index, sf_index in enumerate(sf.index, start=2):
+                    for row_index, sf_index in enumerate(sf.data_df.index, start=2):
                         sf_index.style = get_style_object(row=row_index, column=col_index_in_excel)
                     col_index_in_excel += 1  # Move next to excel indices column
 
                 sf.columns[col_index].style = get_style_object(row=1, column=col_index_in_excel)
-                for row_index, sf_index in enumerate(sf.index, start=start_row_index):
+                for row_index, sf_index in enumerate(sf.data_df.index, start=start_row_index):
                     sf.at[sf_index, col_name].style = get_style_object(row=row_index, column=col_index_in_excel)
                     sf._rows_height[row_index] = sheet.row_dimensions[row_index].height
 
                 sf._columns_width[col_name] = sheet.column_dimensions[sf._get_column_as_letter(sheet, col_name)].width
 
-        sheet_name = kwargs.pop('sheetname', sheet_name)
         header_arg = kwargs.get('header', 0)
         if read_style and isinstance(header_arg, Iterable):
             raise ValueError('Not supporting multiple index columns with read style.')
@@ -250,7 +257,7 @@ class StyleFrame(object):
         return sf
 
     @classmethod
-    def read_excel_as_template(cls, path, df, use_df_boundaries=False, **kwargs):
+    def read_excel_as_template(cls, path: str, df: pd.DataFrame, use_df_boundaries: bool = False, **kwargs) -> 'StyleFrame':
         """
         .. versionadded:: 3.0.1
 
@@ -270,7 +277,7 @@ class StyleFrame(object):
         sf = cls.read_excel(path=path, read_style=True, **kwargs)
 
         num_of_rows, num_of_cols = len(df.index), len(df.columns)
-        template_num_of_rows, template_num_of_cols = len(sf.index), len(sf.columns)
+        template_num_of_rows, template_num_of_cols = len(sf.data_df.index), len(sf.columns)
 
         num_of_cols_to_copy_with_style = min(num_of_cols, template_num_of_cols)
         num_of_rows_to_copy_with_style = min(num_of_rows, template_num_of_rows)
@@ -332,9 +339,24 @@ class StyleFrame(object):
 
         return tuple(range(1, len(self) + 2))
 
-    def to_excel(self, excel_writer='output.xlsx', sheet_name='Sheet1',
-                 allow_protection=False, right_to_left=False, columns_to_hide=None, row_to_add_filters=None,
-                 columns_and_rows_to_freeze=None, best_fit=None, **kwargs):
+    @staticmethod
+    def _to_excel_pandas_defaults(kwargs: Dict[str, Any]) -> Tuple[bool, int, int, str]:
+        """Returns the provided or default values for specific arguments as set by :meth:`pandas.DataFrame.to_excel`
+        """
+
+        header = kwargs.pop('header', True)
+        startcol = kwargs.pop('startcol', 0)
+        startrow = kwargs.pop('startrow', 0)
+        na_rep = kwargs.pop('na_rep', '')
+
+        return header, startcol, startrow, na_rep
+
+    def to_excel(self, excel_writer: Union[str, pd.ExcelWriter, pathlib.Path] = 'output.xlsx',
+                 sheet_name: str = 'Sheet1', allow_protection: bool = False, right_to_left: bool = False,
+                 columns_to_hide: Union[None, str, list, tuple, set] = None, row_to_add_filters: Optional[int] = None,
+                 columns_and_rows_to_freeze: Optional[str] = None, best_fit: Union[None, str, list, tuple, set] = None,
+                 index: bool = False,
+                 **kwargs) -> pd.ExcelWriter:
         """Saves the dataframe to excel and applies the styles.
 
         .. note:: :meth:`to_excel` also accepts all arguments that :meth:`pandas.DataFrame.to_excel` accepts as kwargs.
@@ -369,17 +391,19 @@ class StyleFrame(object):
                       calling ``StyleFrame.to_excel`` by directly modifying ``StyleFrame.A_FACTOR`` and ``StyleFrame.P_FACTOR``
 
         :type best_fit: None or str or list or tuple or set
-        :return: self
-        :rtype: :class:`StyleFrame`
+
+        .. versionadded:: 4.2
+
+        :param bool index: Write row names.
+        :rtype: :class:`pandas.ExcelWriter`
 
         """
 
-        # dealing with needed pandas.to_excel defaults
-        header = kwargs.pop('header', True)
-        index = kwargs.pop('index', False)
-        startcol = kwargs.pop('startcol', 0)
-        startrow = kwargs.pop('startrow', 0)
-        na_rep = kwargs.pop('na_rep', '')
+        if isinstance(excel_writer, pd.ExcelWriter):
+            if excel_writer.engine != 'openpyxl':
+                raise TypeError('styleframe supports only openpyxl, attempted to use {}'.format(excel_writer.engine))
+
+        header, startcol, startrow, na_rep = self._to_excel_pandas_defaults(kwargs)
 
         def get_values(x):
             if isinstance(x, Container):
@@ -393,7 +417,7 @@ class StyleFrame(object):
                 except TypeError:
                     return x
 
-        def within_sheet_boundaries(row=1, column='A'):
+        def within_sheet_boundaries(row: Union[int, str] = 1, column: str = 'A'):
             return (1 <= int(row) <= sheet.max_row
                         and
                     1 <= cell.column_index_from_string(column) <= sheet.max_column)
@@ -417,7 +441,7 @@ class StyleFrame(object):
                                                                                 end_index=end_index)
 
         if len(self.data_df) > 0:
-            export_df = self.data_df.applymap(get_values)
+            export_df = self.data_df.map(get_values)
 
         else:
             export_df = deepcopy(self.data_df)
@@ -525,8 +549,12 @@ class StyleFrame(object):
         if best_fit:
             if not isinstance(best_fit, (list, set, tuple)):
                 best_fit = [best_fit]
-            self.set_column_width_dict({column: (max(self.data_df[column].astype(str).str.len()) + self.A_FACTOR) * self.P_FACTOR
-                                        for column in best_fit})
+            self.set_column_width_dict(
+                {
+                    column: (max(self.data_df[column].astype(str).str.len(), default=0) + self.A_FACTOR) * self.P_FACTOR
+                    for column in best_fit
+                }
+            )
 
         for column in self._columns_width:
             column_letter = self._get_column_as_letter(sheet, column, startcol)
@@ -575,8 +603,14 @@ class StyleFrame(object):
 
         return excel_writer
 
-    def apply_style_by_indexes(self, indexes_to_style, styler_obj, cols_to_style=None, height=None,
-                               complement_style=None, complement_height=None, overwrite_default_style=True):
+    def apply_style_by_indexes(self,
+                               indexes_to_style: Union[list, tuple, int, Container],
+                               styler_obj: Styler,
+                               cols_to_style: Optional[Union[str, Union[List[str], Tuple[str], Set[str]]]] = None,
+                               height: Optional[Union[int, float]] = None,
+                               complement_style: Optional[Styler] = None,
+                               complement_height: Optional[Union[int, float]] = None,
+                               overwrite_default_style: bool = True):
         """
         Applies a certain style to the provided indexes in the dataframe in the provided columns
 
@@ -585,7 +619,7 @@ class StyleFrame(object):
 
             ::
 
-                sf[sf['some_col'] = 20]
+                sf[sf['some_col'] == 20]
 
         :type indexes_to_style: list or tuple or int or Container
         :param styler_obj: `Styler` object that contains the style that will be applied to indexes in `indexes_to_style`
@@ -617,7 +651,7 @@ class StyleFrame(object):
             raise TypeError('styler_obj must be {}, got {} instead.'.format(Styler.__name__, type(styler_obj).__name__))
 
         if isinstance(indexes_to_style, (list, tuple, int)):
-            indexes_to_style = self.index[indexes_to_style]
+            indexes_to_style = self.data_df.index[indexes_to_style]
 
         elif isinstance(indexes_to_style, Container):
             indexes_to_style = pd.Index([indexes_to_style])
@@ -635,21 +669,26 @@ class StyleFrame(object):
         for index in indexes_to_style:
             index.style = style_to_apply
             for col in cols_to_style:
-                self.iloc[self.index.get_loc(index), self.columns.get_loc(col)].style = style_to_apply
+                self.iloc[self.data_df.index.get_loc(index), self.columns.get_loc(col)].style = style_to_apply
 
         if height:
             # Add offset 2 since rows do not include the headers and they starts from 1 (not 0).
-            rows_indexes_for_height_change = [self.index.get_loc(idx) + 2 for idx in indexes_to_style]
+            rows_indexes_for_height_change = [self.data_df.index.get_loc(idx) + 2 for idx in indexes_to_style]
             self.set_row_height(rows=rows_indexes_for_height_change, height=height)
 
         if complement_style:
-            self.apply_style_by_indexes(self.index.difference(indexes_to_style), complement_style, cols_to_style,
+            self.apply_style_by_indexes(self.data_df.index.difference(indexes_to_style), complement_style, cols_to_style,
                                         complement_height if complement_height else height)
 
         return self
 
-    def apply_column_style(self, cols_to_style, styler_obj, style_header=False, use_default_formats=True, width=None,
-                           overwrite_default_style=True):
+    def apply_column_style(self,
+                           cols_to_style: Union[str, List[str], Tuple[str], Set[str]],
+                           styler_obj: Styler,
+                           style_header: bool = False,
+                           use_default_formats: bool = True,
+                           width: Optional[Union[int, float]] = None,
+                           overwrite_default_style: bool = True):
         """Apply style to a whole column
 
         :param cols_to_style: The column names to style.
@@ -684,7 +723,7 @@ class StyleFrame(object):
             if style_header:
                 self.columns[self.columns.get_loc(col_name)].style = style_to_apply
                 self._has_custom_headers_style = True
-            for index in self.index:
+            for index in self.data_df.index:
                 if use_default_formats:
                     if isinstance(self.at[index, col_name].value, pd_timestamp):
                         style_to_apply.number_format = utils.number_formats.date_time
@@ -700,7 +739,10 @@ class StyleFrame(object):
 
         return self
 
-    def apply_headers_style(self, styler_obj, style_index_header=True, cols_to_style=None):
+    def apply_headers_style(self,
+                            styler_obj: Styler,
+                            style_index_header: bool = True,
+                            cols_to_style: Optional[Union[str, List[str], Tuple[str], Set[str]]] = None):
         """Apply style to the headers only
 
         :param styler_obj: The style to apply
@@ -737,7 +779,8 @@ class StyleFrame(object):
         self._has_custom_headers_style = True
         return self
 
-    def set_column_width(self, columns, width):
+    def set_column_width(self, columns: Union[str, Union[str, List[str], Tuple[str], List[int], Tuple[int]]],
+                         width: Union[int, float]) -> 'StyleFrame':
         """Set the width of the given columns
 
         :param columns: Column name(s) or index(es).
@@ -765,7 +808,7 @@ class StyleFrame(object):
 
         return self
 
-    def set_column_width_dict(self, col_width_dict):
+    def set_column_width_dict(self, col_width_dict: Dict[str, Union[int, float]]) -> 'StyleFrame':
         """
         :param col_width_dict: A dictionary from column names to width.
         :type col_width_dict: dict[str, int or float]
@@ -780,7 +823,7 @@ class StyleFrame(object):
 
         return self
 
-    def set_row_height(self, rows, height):
+    def set_row_height(self, rows: Union[int, List[int], Tuple[int], Set[int]], height: Union[int, float]) -> 'StyleFrame':
         """ Set the height of the given rows
 
         :param rows: Row(s) index.
@@ -810,7 +853,7 @@ class StyleFrame(object):
 
         return self
 
-    def set_row_height_dict(self, row_height_dict):
+    def set_row_height_dict(self, row_height_dict: Dict[int, Union[int, float]]) -> 'StyleFrame':
         """
         :param row_height_dict: A dictionary from row indexes to height.
         :type row_height_dict: dict[int, int or float]
@@ -848,7 +891,7 @@ class StyleFrame(object):
                                   if old_col_name in sf._columns_width})
         return sf
 
-    def style_alternate_rows(self, styles, **kwargs):
+    def style_alternate_rows(self, styles: Union[List[Styler], Tuple[Styler]], **kwargs) -> 'StyleFrame':
         """
         .. versionadded:: 1.2
 
@@ -856,21 +899,30 @@ class StyleFrame(object):
 
         .. note:: :meth:`style_alternate_rows` also accepts all arguments that :meth:`apply_style_by_indexes` accepts as kwargs.
 
-        :param styles: List, tuple or set of :class:`.Styler` objects to be applied to rows in an alternating manner
-        :type styles: list[:class:`.Styler`] or tuple[:class:`.Styler`] or set[:class:`.Styler`]
+        :param styles: List or tuple of :class:`.Styler` objects to be applied to rows in an alternating manner
+        :type styles: list[:class:`.Styler`] or tuple[:class:`.Styler`]
         :return: self
         :rtype: :class:`StyleFrame`
 
         """
 
         num_of_styles = len(styles)
-        split_indexes = (self.index[i::num_of_styles] for i in range(num_of_styles))
+        split_indexes = (self.data_df.index[i::num_of_styles] for i in range(num_of_styles))
         for i, indexes in enumerate(split_indexes):
             self.apply_style_by_indexes(indexes, styles[i], **kwargs)
         return self
 
-    def add_color_scale_conditional_formatting(self, start_type, start_value, start_color, end_type, end_value, end_color,
-                                               mid_type=None, mid_value=None, mid_color=None, columns_range=None):
+    def add_color_scale_conditional_formatting(self,
+                                               start_type: str,
+                                               start_value: Union[int, float],
+                                               start_color: str,
+                                               end_type: str,
+                                               end_value: Union[int, float],
+                                               end_color: str,
+                                               mid_type: Optional[str] = None,
+                                               mid_value: Optional[Union[int, float]] = None,
+                                               mid_color: Optional[str] = None,
+                                               columns_range=None):
         """
         :param start_type: The type for the minimum bound
         :type start_type: str: one of :class:`.utils.conditional_formatting_types` or any other type Excel supports