+++ styleframe/styler.py
@@ -1,3 +1,5 @@
+from openpyxl.cell import Cell
+
 from . import utils
 from colour import Color
 from openpyxl.formatting.rule import ColorScaleRule
@@ -5,8 +7,10 @@ from openpyxl.styles import PatternFill, NamedStyle, Color as OpenPyColor, Borde
 from openpyxl.comments import Comment
 from pprint import pformat
 
+from typing import Dict, List, Optional, Union, Set
 
-class Styler(object):
+
+class Styler:
     """
     Used to represent a style
 
@@ -23,8 +27,20 @@ class Styler(object):
     :param bool protection: If ``True``, the cell/column will be write-protected
     :param underline: The underline type
     :type underline: str: one of :class:`.utils.underline` or any other underline Excel supports
-    :param border_type: The border type
-    :type border_type: str: one of :class:`.utils.borders` or any other border type Excel supports
+
+    .. versionchanged:: 4.2
+
+    :param border_type:
+        - If provided a string (one of :class:`.utils.borders` or any other border type Excel supports): all borders
+          will be set to that type.
+        - If provided a set of strings (:class:`.utils.border_locations` or any other border location Excel supports):
+          each provided border will be set to the default border type.
+        - If provided a dict (from location,
+          one of :class:`.utils.border_locations` or any other border location Excel supports) to border type
+          (one of :class:`.utils.borders` or any other border type Excel supports): each provided border will be set to
+          the provided border type.
+
+    :type border_type: str or set[str] or dict[str, str]
 
     .. versionadded:: 1.2
 
@@ -55,33 +71,46 @@ class Styler(object):
 
     .. note:: For any of ``date_format``, ``time_format`` and ``date_time_format`` to take effect, the value being
               styled must be an actual ``date``/``time``/``datetime`` object.
+
+    .. versionadded:: 4.1
+
+    :param bool strikethrough:
+    :param bool italic:
     """
 
-    cache = {}
+    cache: Dict['Styler', NamedStyle] = {}
 
-    def __init__(self, bg_color=None, bold=False, font=utils.fonts.arial, font_size=12.0, font_color=None,
-                 number_format=utils.number_formats.general, protection=False, underline=None,
-                 border_type=utils.borders.thin, horizontal_alignment=utils.horizontal_alignments.center,
-                 vertical_alignment=utils.vertical_alignments.center, wrap_text=True, shrink_to_fit=True,
-                 fill_pattern_type=utils.fill_pattern_types.solid, indent=0.0, comment_author=None, comment_text=None,
-                 text_rotation=0, date_format=utils.number_formats.date,
-                 time_format=utils.number_formats.time_24_hours, date_time_format=utils.number_formats.date_time):
+    def __init__(self,
+                 bg_color: Optional[str] = None,
+                 bold: bool = False,
+                 font: str = utils.fonts.arial,
+                 font_size: Union[int, float] = 12.0,
+                 font_color: Optional[str] = None,
+                 number_format: str = utils.number_formats.general,
+                 protection: bool = False,
+                 underline: Optional[str] = None,
+                 border_type: Union[str, Set[str], Dict[str, str]] = utils.borders.thin,
+                 horizontal_alignment: str = utils.horizontal_alignments.center,
+                 vertical_alignment: str = utils.vertical_alignments.center,
+                 wrap_text: bool = True,
+                 shrink_to_fit: bool = True,
+                 fill_pattern_type: str = utils.fill_pattern_types.solid,
+                 indent: Union[int, float] = 0.0,
+                 comment_author: Optional[str] = None,
+                 comment_text: Optional[str] = None,
+                 text_rotation: int = 0,
+                 date_format: str = utils.number_formats.date,
+                 time_format: str = utils.number_formats.time_24_hours,
+                 date_time_format: str = utils.number_formats.date_time,
+                 strikethrough: bool = False,
+                 italic: bool = False):
 
-        def get_color_from_string(color_str, default_color=None):
+        def get_color_from_string(color_str: str, default_color: Optional[str] = None) -> str:
             if color_str and color_str.startswith('#'):
                 color_str = color_str[1:]
             if not utils.is_hex_color_string(hex_string=color_str):
                 color_str = utils.colors.get(color_str, default_color)
             return color_str
-
-        if border_type == utils.borders.default_grid:
-            if bg_color is not None or fill_pattern_type != utils.fill_pattern_types.solid:
-                raise ValueError('`bg_color`or `fill_pattern_type` conflict with border_type={}'.format(utils.borders.default_grid))
-            self.border_type = None
-            self.fill_pattern_type = None
-        else:
-            self.border_type = border_type
-            self.fill_pattern_type = fill_pattern_type
 
         self.bold = bold
         self.font = font
@@ -102,14 +131,34 @@ class Styler(object):
         self.date_format = date_format
         self.time_format = time_format
         self.date_time_format = date_time_format
+        self.strikethrough = strikethrough
+        self.italic = italic
+        self.fill_pattern_type = fill_pattern_type
+
+        if isinstance(border_type, set):
+            self.border_type = {border_location: utils.borders.thin for border_location in border_type}
+        elif isinstance(border_type, dict):
+            self.border_type = border_type
+        else:
+            self.border_type = {
+                utils.border_locations.top: border_type,
+                utils.border_locations.right: border_type,
+                utils.border_locations.bottom: border_type,
+                utils.border_locations.left: border_type
+            }
+
+        if border_type == utils.borders.default_grid:
+            if bg_color is not None or fill_pattern_type != utils.fill_pattern_types.solid:
+                raise ValueError(f'`bg_color`or `fill_pattern_type` conflict with border_type={utils.borders.default_grid}')
+            self.border_type = None
+            self.fill_pattern_type = None
 
     def __eq__(self, other):
-        if not isinstance(other, self.__class__):
-            return False
-        return self.__dict__ == other.__dict__
+        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
 
     def __hash__(self):
-        return hash(tuple((k, v) for k, v in sorted(self.__dict__.items())))
+        return hash(tuple((k, v) if not isinstance(v, (dict, set)) else hash(tuple(v.items()))
+                          for k, v in sorted(self.__dict__.items())))
 
     def __add__(self, other):
         default = Styler().__dict__
@@ -135,12 +184,18 @@ class Styler(object):
         try:
             openpyxl_style = self.cache[self]
         except KeyError:
-            side = Side(border_style=self.border_type, color=utils.colors.black)
-            border = Border(left=side, right=side, top=side, bottom=side)
+            if isinstance(self.border_type, str):
+                side = Side(border_style=self.border_type, color=utils.colors.black)
+                border = Border(left=side, right=side, top=side, bottom=side)
+            else:
+                border = Border(**{border_location: Side(border_style=border_type, color=utils.colors.black)
+                                   for border_location, border_type in self.border_type.items()})
+
             openpyxl_style = self.cache[self] = NamedStyle(
                 name=str(hash(self)),
                 font=Font(name=self.font, size=self.font_size, color=OpenPyColor(self.font_color),
-                          bold=self.bold, underline=self.underline),
+                          bold=self.bold, underline=self.underline, strikethrough=self.strikethrough,
+                          italic=self.italic),
                 fill=PatternFill(patternType=self.fill_pattern_type, fgColor=self.bg_color),
                 alignment=Alignment(horizontal=self.horizontal_alignment, vertical=self.vertical_alignment,
                                     wrap_text=self.wrap_text, shrink_to_fit=self.shrink_to_fit,
@@ -152,7 +207,9 @@ class Styler(object):
         return openpyxl_style
 
     @classmethod
-    def from_openpyxl_style(cls, openpyxl_style, theme_colors, openpyxl_comment=None):
+    def from_openpyxl_style(cls, openpyxl_style: Cell, theme_colors: List[str],
+                            openpyxl_comment: Optional[Comment] = None):
+
         def _calc_new_hex_from_theme_hex_and_tint(theme_hex, color_tint):
             if not theme_hex.startswith('#'):
                 theme_hex = '#' + theme_hex
@@ -160,11 +217,19 @@ class Styler(object):
             color_obj.luminance = _calc_lum_from_tint(color_tint, color_obj.luminance)
             return color_obj.hex_l[1:]
 
-        def _calc_lum_from_tint(color_tint, current_lum):
-            # based on http://ciintelligence.blogspot.co.il/2012/02/converting-excel-theme-color-and-tint.html
-            if not color_tint:
+        def _calc_lum_from_tint(color_tint: Optional[float], current_lum: float) -> float:
+            """"
+            Based on https://ciintelligence.blogspot.co.il/2012/02/converting-excel-theme-color-and-tint.html
+            """
+            if color_tint is None:
                 return current_lum
-            return current_lum * (1.0 + color_tint)
+
+            current_lum *= 255
+
+            if color_tint < 0:
+                return current_lum * (1.0 + color_tint) / 255
+
+            return (current_lum * (1.0 - color_tint) + (255 - 255 * (1.0 - color_tint))) / 255
 
         bg_color = openpyxl_style.fill.fgColor.rgb
 
@@ -178,6 +243,8 @@ class Styler(object):
             bg_color = _calc_new_hex_from_theme_hex_and_tint(bg_color, tint)
 
         bold = openpyxl_style.font.bold
+        strikethrough = openpyxl_style.font.strikethrough
+        italic = openpyxl_style.font.italic
         font = openpyxl_style.font.name
         font_size = openpyxl_style.font.size
         try:
@@ -197,7 +264,6 @@ class Styler(object):
         number_format = openpyxl_style.number_format
         protection = openpyxl_style.protection.locked
         underline = openpyxl_style.font.underline
-        border_type = openpyxl_style.border.bottom.border_style
         horizontal_alignment = openpyxl_style.alignment.horizontal
         vertical_alignment = openpyxl_style.alignment.vertical
         wrap_text = openpyxl_style.alignment.wrap_text or False
@@ -205,6 +271,13 @@ class Styler(object):
         fill_pattern_type = openpyxl_style.fill.patternType
         indent = openpyxl_style.alignment.indent
         text_rotation = openpyxl_style.alignment.text_rotation
+
+        border_type = {
+            utils.border_locations.top: getattr(openpyxl_style.border.top, 'border_style', None),
+            utils.border_locations.right: getattr(openpyxl_style.border.right, 'border_style', None),
+            utils.border_locations.bottom: getattr(openpyxl_style.border.bottom, 'border_style', None),
+            utils.border_locations.left: getattr(openpyxl_style.border.left, 'border_style', None)
+        }
 
         if openpyxl_comment:
             comment_author = openpyxl_comment.author
@@ -217,10 +290,11 @@ class Styler(object):
                    number_format, protection, underline,
                    border_type, horizontal_alignment,
                    vertical_alignment, wrap_text, shrink_to_fit,
-                   fill_pattern_type, indent, comment_author, comment_text, text_rotation)
+                   fill_pattern_type, indent, comment_author, comment_text, text_rotation,
+                   strikethrough=strikethrough, italic=italic)
 
     @classmethod
-    def combine(cls, *styles):
+    def combine(cls, *styles: 'Styler'):
         """
         .. versionadded:: 1.6
 
@@ -248,7 +322,7 @@ class Styler(object):
     create_style = to_openpyxl_style
 
 
-class ColorScaleConditionalFormatRule(object):
+class ColorScaleConditionalFormatRule:
     """Creates a color scale conditional format rule. Wraps openpyxl's ColorScaleRule.
     Mostly should not be used directly, but through StyleFrame.add_color_scale_conditional_formatting
     """